Move RunnerDNJ to dnj script
